use :oauth-token instead of :auth
The way the README is currently written made it sound like you just put :oauth-token in to the options map _in addition to_ :auth. When you add it, the header that gets put into the request is

    >> Authorization: Basic asfdhjkfshjdfsdhjsfhdja897243879487943=

So it's still trying to do basic auth with the username/password. You really need to use :oauth-token _instead_ of :auth. When you replace it, the header that gets put into the request is

    >> Authorization: token 87934589534980dgsflnjkldfgsfgsfd

For me this was a "Personal access token" created from the [Applications Settings](https://github.com/settings/applications) page.